Streamlined the package to just the SewerRat interface.
All functions now expect a url= argument, as any default setting for the
URL is the responsibility of the application that uses this package.
The restUrl() function and .onLoad hook have been removed.

We also remove the quickSave/Load functions, which don't have much to do
with the SewerRat API and should be implemented elsewhere.